LPS-89175 ReplaceJournalExportImportContentReferences test added

+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-test/src/testIntegration/java/com/liferay/dynamic/data/mapping/exportimport/content/processor/test/DDMFormValuesExportImportContentProcessorTest.java
@@ -16,14 +16,19 @@ package com.liferay.dynamic.data.mapping.exportimport.content.processor.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.document.library.kernel.model.DLFileEntry;
+import com.liferay.document.library.kernel.model.DLFileEntryMetadata;
+import com.liferay.document.library.kernel.model.DLFileEntryType;
 import com.liferay.document.library.kernel.model.DLFolderConstants;
-import com.liferay.document.library.kernel.service.DLAppLocalServiceUtil;
+import com.liferay.document.library.kernel.service.DLAppLocalService;
 import com.liferay.document.library.kernel.service.DLFileEntryLocalService;
+import com.liferay.document.library.kernel.service.DLFileEntryTypeLocalService;
 import com.liferay.dynamic.data.mapping.helper.DDMFormInstanceTestHelper;
 import com.liferay.dynamic.data.mapping.model.DDMForm;
+import com.liferay.dynamic.data.mapping.model.DDMFormField;
 import com.liferay.dynamic.data.mapping.model.DDMFormInstance;
 import com.liferay.dynamic.data.mapping.model.DDMStructure;
 import com.liferay.dynamic.data.mapping.model.DDMTemplate;
+import com.liferay.dynamic.data.mapping.model.LocalizedValue;
 import com.liferay.dynamic.data.mapping.model.Value;
 import com.liferay.dynamic.data.mapping.service.DDMStructureLocalService;
 import com.liferay.dynamic.data.mapping.service.DDMTemplateLocalService;
@@ -31,6 +36,7 @@ import com.liferay.dynamic.data.mapping.storage.DDMFormFieldValue;
 import com.liferay.dynamic.data.mapping.storage.DDMFormValues;
 import com.liferay.dynamic.data.mapping.test.util.DDMStructureTestUtil;
 import com.liferay.dynamic.data.mapping.test.util.DDMTemplateTestUtil;
+import com.liferay.dynamic.data.mapping.util.DDMBeanTranslatorUtil;
 import com.liferay.exportimport.content.processor.ExportImportContentProcessor;
 import com.liferay.exportimport.kernel.lar.PortletDataContext;
 import com.liferay.exportimport.kernel.lar.PortletDataContextFactoryUtil;
@@ -39,7 +45,10 @@ import com.liferay.exportimport.test.util.TestUserIdStrategy;
 import com.liferay.journal.model.JournalArticle;
 import com.liferay.journal.model.JournalFolderConstants;
 import com.liferay.journal.service.JournalArticleLocalService;
+import com.liferay.journal.test.util.JournalTestUtil;
 import com.liferay.petra.string.StringBundler;
+import com.liferay.petra.string.StringPool;
+import com.liferay.portal.kernel.io.unsync.UnsyncByteArrayInputStream;
 import com.liferay.portal.kernel.json.JSONFactoryUtil;
 import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.model.Group;
@@ -54,7 +63,9 @@ import com.liferay.portal.kernel.test.util.ServiceContextTestUtil;
 import com.liferay.portal.kernel.test.util.TestDataConstants;
 import com.liferay.portal.kernel.test.util.TestPropsValues;
 import com.liferay.portal.kernel.util.ContentTypes;
+import com.liferay.portal.kernel.util.FileUtil;
 import com.liferay.portal.kernel.util.LocaleUtil;
+import com.liferay.portal.kernel.util.MimeTypesUtil;
 import com.liferay.portal.kernel.util.Time;
 import com.liferay.portal.kernel.xml.Document;
 import com.liferay.portal.kernel.xml.Element;
@@ -67,11 +78,16 @@ import com.liferay.registry.Registry;
 import com.liferay.registry.RegistryUtil;
 import com.liferay.registry.ServiceTracker;
 
+import java.io.File;
+import java.io.InputStream;
+
 import java.util.Date;
 import java.util.HashMap;
+import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
+import java.util.Set;
 
 import org.junit.After;
 import org.junit.Assert;
@@ -122,10 +138,6 @@ public class DDMFormValuesExportImportContentProcessorTest {
 
 		_stagingGroup = _liveGroup.getStagingGroup();
 
-		ServiceContext serviceContext =
-			ServiceContextTestUtil.getServiceContext(
-				_stagingGroup.getGroupId(), TestPropsValues.getUserId());
-
 		TestReaderWriter testReaderWriter = new TestReaderWriter();
 
 		_portletDataContextExport =
@@ -167,80 +179,27 @@ public class DDMFormValuesExportImportContentProcessorTest {
 
 		rootElement.addElement("entry");
 
-		_fileEntry = DLAppLocalServiceUtil.addFileEntry(
-			TestPropsValues.getUserId(), _stagingGroup.getGroupId(),
-			DLFolderConstants.DEFAULT_PARENT_FOLDER_ID,
-			RandomTestUtil.randomString() + ".txt", ContentTypes.TEXT_PLAIN,
-			TestDataConstants.TEST_BYTE_ARRAY, serviceContext);
-
-		ThumbnailCapability thumbnailCapability =
-			_fileEntry.getRepositoryCapability(ThumbnailCapability.class);
-
-		_fileEntry = thumbnailCapability.setLargeImageId(
-			_fileEntry, _fileEntry.getFileEntryId());
-
-		_formInstance = createFormInstanceWithDocLib(_stagingGroup, _fileEntry);
-
-		long classNameId = ClassNameLocalServiceUtil.getClassNameId(
-			JournalArticle.class);
-
-		DDMStructure structure = _formInstance.getStructure();
-
-		DDMForm ddmForm = _formInstance.getDDMForm();
-
-		structure.setDDMForm(ddmForm);
-
-		structure.setClassNameId(classNameId);
-
-		_ddmStructureLocalService.updateDDMStructure(structure);
-
-		_ddmTemplate = DDMTemplateTestUtil.addTemplate(
-			_stagingGroup.getGroupId(), structure.getStructureId(),
-			classNameId);
-
-		Map<Locale, String> nameMap = new HashMap<>();
-
-		nameMap.put(LocaleUtil.US, RandomTestUtil.randomString());
-
-		StringBundler sb = new StringBundler(16);
-
-		sb.append("<?xml version=\"1.0\"?>  ");
-		sb.append("<root available-locales=\"en_US\" ");
-		sb.append("default-locale=\"en_US\"> \t");
-		sb.append("<dynamic-element name=\"DocumentsAndMedia9t17\" ");
-		sb.append("type=\"document_library\" index-type=\"keyword\" ");
-		sb.append("instance-id=\"lvsi\"> \t\t");
-		sb.append("<dynamic-content language-id=\"en_US\">");
-		sb.append("<![CDATA[{\"classPK\":\"");
-		sb.append(_fileEntry.getFileEntryId());
-		sb.append("\",\"groupId\":\"");
-		sb.append(_fileEntry.getGroupId());
-		sb.append("\",\"title\":\"");
-		sb.append(_fileEntry.getTitle());
-		sb.append("\",\"type\":\"document\",\"uuid\":\"");
-		sb.append(_fileEntry.getUuid());
-		sb.append("\"}]]></dynamic-content> \t</dynamic-element> </root>");
-
-		String content = sb.toString();
-
-		_journalArticle = _journalArticleLocalService.addArticle(
-			TestPropsValues.getUserId(), _stagingGroup.getGroupId(),
-			JournalFolderConstants.DEFAULT_PARENT_FOLDER_ID, nameMap, nameMap,
-			content, structure.getStructureKey(), _ddmTemplate.getTemplateKey(),
-			serviceContext);
-
 		_exportImportContentProcessor = _serviceTracker.getService();
 	}
 
 	@After
 	public void tearDown() throws Exception {
-		_journalArticleLocalService.deleteArticle(_journalArticle);
-		_ddmTemplateLocalService.deleteDDMTemplate(_ddmTemplate);
-		_ddmStructureLocalService.deleteDDMStructure(_ddmStructure);
+		_journalArticleLocalService.deleteArticles(_stagingGroup.getGroupId());
+		_journalArticleLocalService.deleteArticles(_liveGroup.getGroupId());
+
+		if (_ddmTemplate != null) {
+			_ddmTemplateLocalService.deleteDDMTemplate(_ddmTemplate);
+		}
+
+		if (_ddmStructure != null) {
+			_ddmStructureLocalService.deleteDDMStructure(_ddmStructure);
+		}
 	}
 
 	@Test
-	public void testReplaceExportImportContentReferences() throws Exception {
+	public void testReplaceDLExportImportContentReferences() throws Exception {
+		_initDLReferences();
+
 		DDMFormValues settingsDDMFormValues =
 			_formInstance.getSettingsDDMFormValues();
 
@@ -303,6 +262,62 @@ public class DDMFormValuesExportImportContentProcessorTest {
 		Assert.assertEquals(newDLFileEntryId, jsonObject.getLong("classPK"));
 	}
 
+	@Test
+	public void testReplaceJournalExportImportContentReferences()
+		throws Exception {
+
+		_initJournalReferences();
+
+		DDMFormValues exportDDMFormValues =
+			_exportImportContentProcessor.replaceExportContentReferences(
+				_portletDataContextExport, _fileEntry, _journalDDMFormValues,
+				true, true);
+
+		Map<Long, Long> groupIds =
+			(Map<Long, Long>)_portletDataContextImport.getNewPrimaryKeysMap(
+				Group.class);
+
+		groupIds.put(_stagingGroup.getGroupId(), _liveGroup.getGroupId());
+
+		Map<Long, Long> classPKs =
+			(Map<Long, Long>)_portletDataContextImport.getNewPrimaryKeysMap(
+				JournalArticle.class);
+
+		Map<Locale, String> nameMap = new HashMap<>();
+
+		nameMap.put(LocaleUtil.US, RandomTestUtil.randomString());
+
+		long resourcePrimKey = _journalArticle.getResourcePrimKey();
+
+		JournalArticle newJournalArticle = JournalTestUtil.addArticle(
+			TestPropsValues.getUserId(), _liveGroup.getGroupId(),
+			DLFolderConstants.DEFAULT_PARENT_FOLDER_ID);
+
+		newJournalArticle.setUuid(_journalArticle.getUuid());
+
+		_journalArticleLocalService.updateJournalArticle(newJournalArticle);
+
+		classPKs.put(resourcePrimKey, newJournalArticle.getResourcePrimKey());
+
+		_exportImportContentProcessor.replaceImportContentReferences(
+			_portletDataContextImport, _fileEntry, exportDDMFormValues);
+
+		List<DDMFormFieldValue> ddmFormFieldValues =
+			exportDDMFormValues.getDDMFormFieldValues();
+
+		DDMFormFieldValue ddmFormFieldValue = ddmFormFieldValues.get(0);
+
+		Value value = ddmFormFieldValue.getValue();
+
+		JSONObject jsonObject = JSONFactoryUtil.createJSONObject(
+			value.getString(LocaleUtil.US));
+
+		long newArticleResourcePrimKey = newJournalArticle.getResourcePrimKey();
+
+		Assert.assertEquals(
+			newArticleResourcePrimKey, jsonObject.getLong("classPK"));
+	}
+
 	protected DDMFormInstance createFormInstanceWithDocLib(
 			Group group, FileEntry fileEntry)
 		throws Exception {
@@ -315,6 +330,199 @@ public class DDMFormValuesExportImportContentProcessorTest {
 
 		return ddmFormInstanceTestHelper.addDDMFormInstance(
 			_ddmStructure, fileEntry);
+	}
+
+	private DDMForm _createDDMFormWithJournalField(
+			Group group, JournalArticle journalArticle)
+		throws Exception {
+
+		_ddmStructure = DDMStructureTestUtil.addStructure(
+			group.getGroupId(), DLFileEntryMetadata.class.getName());
+
+		Map<Locale, String> nameMap = new HashMap<>();
+
+		nameMap.put(LocaleUtil.US, RandomTestUtil.randomString());
+
+		Map<Locale, String> descriptionMap = new HashMap<>();
+
+		descriptionMap.put(LocaleUtil.US, RandomTestUtil.randomString());
+
+		DDMForm journalDDMForm = new DDMForm();
+
+		Set<Locale> availableLocales = new LinkedHashSet<>();
+
+		availableLocales.add(LocaleUtil.US);
+
+		journalDDMForm.setAvailableLocales(availableLocales);
+
+		journalDDMForm.setDefaultLocale(LocaleUtil.US);
+
+		List<DDMFormField> ddmFormFields = journalDDMForm.getDDMFormFields();
+
+		DDMFormField webContentFormField = new DDMFormField(
+			"WebContenttest", "ddm-journal-article");
+
+		webContentFormField.setDataType("journal-article");
+		webContentFormField.setLocalizable(true);
+		webContentFormField.setShowLabel(true);
+		webContentFormField.setDDMForm(journalDDMForm);
+		webContentFormField.setLocalizable(true);
+		webContentFormField.setFieldNamespace("ddm");
+
+		ddmFormFields.add(webContentFormField);
+
+		long classNameId = ClassNameLocalServiceUtil.getClassNameId(
+			DLFileEntryMetadata.class);
+
+		_ddmStructure.setDDMForm(journalDDMForm);
+
+		_ddmStructure.setClassNameId(classNameId);
+
+		_ddmStructureLocalService.updateDDMStructure(_ddmStructure);
+
+		_journalDDMFormValues = new DDMFormValues(journalDDMForm);
+
+		_journalDDMFormValues.setAvailableLocales(availableLocales);
+		_journalDDMFormValues.setDefaultLocale(LocaleUtil.US);
+
+		JSONObject jsonObject = JSONFactoryUtil.createJSONObject();
+
+		jsonObject.put("className", JournalArticle.class.getName());
+		jsonObject.put("classPK", journalArticle.getResourcePrimKey());
+		jsonObject.put("title", journalArticle.getTitle());
+
+		List<DDMFormField> journalFormFields =
+			journalDDMForm.getDDMFormFields();
+
+		for (DDMFormField currentFormField : journalFormFields) {
+			LocalizedValue value = new LocalizedValue();
+
+			value.addString(LocaleUtil.US, jsonObject.toString());
+
+			DDMFormFieldValue ddmFormFieldValue = new DDMFormFieldValue();
+
+			ddmFormFieldValue.setDDMFormValues(_journalDDMFormValues);
+			ddmFormFieldValue.setName(currentFormField.getName());
+			ddmFormFieldValue.setValue(value);
+
+			_journalDDMFormValues.addDDMFormFieldValue(ddmFormFieldValue);
+		}
+
+		return journalDDMForm;
+	}
+
+	private void _initDLReferences() throws Exception {
+		ServiceContext serviceContext =
+			ServiceContextTestUtil.getServiceContext(
+				_stagingGroup.getGroupId(), TestPropsValues.getUserId());
+
+		_fileEntry = _dlAppLocalService.addFileEntry(
+			TestPropsValues.getUserId(), _stagingGroup.getGroupId(),
+			DLFolderConstants.DEFAULT_PARENT_FOLDER_ID,
+			RandomTestUtil.randomString() + ".txt", ContentTypes.TEXT_PLAIN,
+			TestDataConstants.TEST_BYTE_ARRAY, serviceContext);
+
+		ThumbnailCapability thumbnailCapability =
+			_fileEntry.getRepositoryCapability(ThumbnailCapability.class);
+
+		_fileEntry = thumbnailCapability.setLargeImageId(
+			_fileEntry, _fileEntry.getFileEntryId());
+
+		_formInstance = createFormInstanceWithDocLib(_stagingGroup, _fileEntry);
+
+		DDMStructure structure = _formInstance.getStructure();
+
+		long classNameId = ClassNameLocalServiceUtil.getClassNameId(
+			JournalArticle.class);
+
+		DDMForm ddmForm = _formInstance.getDDMForm();
+
+		structure.setDDMForm(ddmForm);
+
+		structure.setClassNameId(classNameId);
+
+		_ddmStructureLocalService.updateDDMStructure(structure);
+
+		_ddmTemplate = DDMTemplateTestUtil.addTemplate(
+			_stagingGroup.getGroupId(), structure.getStructureId(),
+			classNameId);
+
+		Map<Locale, String> nameMap = new HashMap<>();
+
+		nameMap.put(LocaleUtil.US, RandomTestUtil.randomString());
+
+		StringBundler sb = new StringBundler(16);
+
+		sb.append("<?xml version=\"1.0\"?>  ");
+		sb.append("<root available-locales=\"en_US\" ");
+		sb.append("default-locale=\"en_US\"> \t");
+		sb.append("<dynamic-element name=\"DocumentsAndMedia9t17\" ");
+		sb.append("type=\"document_library\" index-type=\"keyword\" ");
+		sb.append("instance-id=\"lvsi\"> \t\t");
+		sb.append("<dynamic-content language-id=\"en_US\">");
+		sb.append("<![CDATA[{\"classPK\":\"");
+		sb.append(_fileEntry.getFileEntryId());
+		sb.append("\",\"groupId\":\"");
+		sb.append(_fileEntry.getGroupId());
+		sb.append("\",\"title\":\"");
+		sb.append(_fileEntry.getTitle());
+		sb.append("\",\"type\":\"document\",\"uuid\":\"");
+		sb.append(_fileEntry.getUuid());
+		sb.append("\"}]]></dynamic-content> \t</dynamic-element> </root>");
+
+		String content = sb.toString();
+
+		_journalArticle = _journalArticleLocalService.addArticle(
+			TestPropsValues.getUserId(), _stagingGroup.getGroupId(),
+			JournalFolderConstants.DEFAULT_PARENT_FOLDER_ID, nameMap, nameMap,
+			content, structure.getStructureKey(), _ddmTemplate.getTemplateKey(),
+			serviceContext);
+	}
+
+	private void _initJournalReferences() throws Exception {
+		ServiceContext serviceContext =
+			ServiceContextTestUtil.getServiceContext(
+				_stagingGroup.getGroupId(), TestPropsValues.getUserId());
+
+		Map<Locale, String> nameMap = new HashMap<>();
+
+		nameMap.put(LocaleUtil.US, RandomTestUtil.randomString());
+
+		_journalArticle = JournalTestUtil.addArticle(
+			TestPropsValues.getUserId(), _stagingGroup.getGroupId(),
+			JournalFolderConstants.DEFAULT_PARENT_FOLDER_ID);
+
+		InputStream is = new UnsyncByteArrayInputStream(new byte[0]);
+		long size = 0;
+		File file = FileUtil.createTempFile(is);
+
+		Map<String, com.liferay.dynamic.data.mapping.kernel.DDMFormValues>
+			ddmFormValuesMap = new HashMap<>();
+
+		_createDDMFormWithJournalField(_stagingGroup, _journalArticle);
+
+		ddmFormValuesMap.put(
+			_ddmStructure.getStructureKey(),
+			DDMBeanTranslatorUtil.translate(_journalDDMFormValues));
+
+		DLFileEntryType dlFileEntryType =
+			_dlFileEntryTypeLocalService.addFileEntryType(
+				TestPropsValues.getUserId(), _stagingGroup.getGroupId(),
+				RandomTestUtil.randomString(), RandomTestUtil.randomString(),
+				new long[] {_ddmStructure.getStructureId()}, serviceContext);
+
+		DLFileEntry dlFileEntry = _dlFileEntryLocalService.addFileEntry(
+			TestPropsValues.getUserId(), _stagingGroup.getGroupId(),
+			_stagingGroup.getGroupId(),
+			DLFolderConstants.DEFAULT_PARENT_FOLDER_ID,
+			RandomTestUtil.randomString() + ".txt",
+			MimeTypesUtil.getExtensionContentType("txt"),
+			RandomTestUtil.randomString(), RandomTestUtil.randomString(),
+			StringPool.BLANK, dlFileEntryType.getFileEntryTypeId(),
+			ddmFormValuesMap, file, is, size, serviceContext);
+
+		_fileEntry = _dlAppLocalService.getFileEntry(
+			dlFileEntry.getFileEntryId());
 	}
 
 	private static ServiceTracker
@@ -332,7 +540,13 @@ public class DDMFormValuesExportImportContentProcessorTest {
 	private DDMTemplateLocalService _ddmTemplateLocalService;
 
 	@Inject
+	private DLAppLocalService _dlAppLocalService;
+
+	@Inject
 	private DLFileEntryLocalService _dlFileEntryLocalService;
+
+	@Inject
+	private DLFileEntryTypeLocalService _dlFileEntryTypeLocalService;
 
 	private ExportImportContentProcessor<DDMFormValues>
 		_exportImportContentProcessor;
@@ -343,6 +557,7 @@ public class DDMFormValuesExportImportContentProcessorTest {
 	@Inject
 	private JournalArticleLocalService _journalArticleLocalService;
 
+	private DDMFormValues _journalDDMFormValues;
 	private Group _liveGroup;
 	private PortletDataContext _portletDataContextExport;
 	private PortletDataContext _portletDataContextImport;
